powerpc: size_t must be a 32 bit value :-/

+++ include/sys/powerpc/_types.h
@@ -52,10 +52,10 @@ typedef uint64_t	size_t;
 /*
  * Below are standard types.
  */
-typedef uint64_t	addr_t;
+typedef uint32_t	addr_t;
 typedef uint64_t	off_t;
-typedef int64_t		ssize_t;
-typedef uint64_t	register_t;
+typedef int32_t		ssize_t;
+typedef uint32_t	register_t;
 typedef int64_t		time_t;
 typedef int32_t		clock_t;
